Tests for ParameterFactory added
